nrunner: escape command line arguments that stat with a dash
Because argparser will attempt to parse arguments that start with a
dash as yet another option, instead of the value to the '-a' option,
we need to escape them.

This is a complement to 48826017e, which introduced this concept and
the decoding of base64 arguments, but failed to encode the necessary
ones when creating a list of arguments.

Signed-off-by: Cleber Rosa <crosa@redhat.com>