utils/software_manager: Make SUPPORTED_PACKAGE_MANAGERS a map
The utils/software_manager.py module has the SUPPORTED_PACKAGE_MANAGERS
global attribute which list the name of all backend implemented, and
should be updated on occasion of a new backend is introduced. However
the backend implementer also needs to update the mapping in
SoftwareManager._init_on_demand(). Instead let's consolidate the list
(and the mapping name <-> backend) in a single place, converting
SUPPORTED_PACKAGE_MANAGERS to a map.

Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>