utils/software_manager: Get rid of SUPPORTED_PACKAGE_MANAGERS
The utils/software_manager.py module has the SUPPORTED_PACKAGE_MANAGERS
global attribute which list the name of all backend implemented, and
should be updated on occasion of a new backend is introduced. However
the backend implementer also needs to update the mapping in
SoftwareManager._init_on_demand(). Instead let's consolidate the list
(and the mapping name <-> backend) in a single place, which is the added
_get_backends() module method.

Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>