python-avocado.spec: ignore Python warnings when running tests
Even though we adjusted some tests to be less strict about the output
produced by Avocado, we currently ignore those on:
  * make check
  * Travis CI

So it seems fair that the other place that runs those tests (during an
RPM build) do the same.

BTW, these "Python warnings" are really tips that Avocado itself gives
to users, for instance, with regards to upcoming deprecation of features.

Signed-off-by: Cleber Rosa <crosa@redhat.com>