avocado/utils/ssh.py: allow access to the SSH raw local command
Some users want to have more flexibility in how they interact with
the ssh process.

Given that most of the value in this module has to do with creating
the command lines that will execute on remote systems, it seems fair
to expose that.  This seems to be one way to allow for many use cases
without bloating the API or introducing much complexity.

Signed-off-by: Cleber Rosa <crosa@redhat.com>