moved --paginator to a global settings
With this new settings model, every option should be inside a
'namespace'. Since that a few plugins are using `--paginator` and
Avocado code was looking for paginator at the settings' root, makes
sense to move to a more global context. This change will move
--paginator to a global option and associate this option with
'core.paginator' in the configuration file.

Signed-off-by: Beraldo Leal <bleal@redhat.com>