Misc Asset: Expose find_asset_file method.
This commit exposes find_asset_file private method and remove the
necessity of a parameter to call it, adding the necessary logic
to this method itself. Also, it removes duplicated logic from
get_metadata method after the changes on find_asset_file.

This is a pre-work for extending the Asset module to allow the
skip of a test if an asset is not available.

Signed-off-by: Willian Rampazzo <willianr@redhat.com>