move --verbose to a global option
With this new settings model, every option should be inside a
'namespace'.  Since that list plugin was using `--verbose` and Avocado
code was looking at the settings' root, this makes sense to move to a
more global context. This change will move --verbose to a global option
and associate this option with 'core.verbose' in the configuration file.

Signed-off-by: Beraldo Leal <bleal@redhat.com>