Extends fetch_asset core method to find assets and cancel the test.
This extension allows the user to choose if the fetch_asset method
should move the asset to the cache or should just look for the
asset in the cache.

When the user wants just to find the asset in the cache, if the
asset is not available, it is possible to cancel the test.

Signed-off-by: Willian Rampazzo <willianr@redhat.com>