N(ext) Runner: make known runners and explicit requirement
Explicit is better than implicit, and for users of the
"avocado.core.nrunner.Task" class (and related utilities), known
runners were being set implicitly with the ones defined in the module.

Due to the nature of the nrunner, that allows other runners to be defined
in other modules, it's better to be explicit about them.

Signed-off-by: Cleber Rosa <crosa@redhat.com>