Extends fetch_asset core method to find assets and cancel the test.
This extension allows the user to choose if the fetch_asset core
method should move the asset to the cache or should just look for
the asset in the cache.

The user can cancel the test when the asset is not available in
the cache or when there was a problem adding the asset to the
cache.

Signed-off-by: Willian Rampazzo <willianr@redhat.com>