avocado/utils/ssh.py: do not capture and supress exception at __exit__
It was never the intention, AFAICT, to capture and supress the
exceptions that could occurr during the use of an ssh.Session()
instance as a context manager.  But, because it was returning True,
that was effectively what was happening.

By not returning True, any exceptions that happen during the context
will be propagated.

Reference: https://docs.python.org/3/reference/datamodel.html#object.__exit__
Signed-off-by: Cleber Rosa <crosa@redhat.com>