utils/network: Big refactoring on network library to improve consistency
Multiple pull requests were trying to add more features to this library.
But because of a lack of convention, it was getting hard to understand
this and we were losing usability. Besides a complete rewrite here, this
pull request is a) adding a complete set of methods to handle network
interfaces parameters; b) adding a wait when setting MTU values on
interfaces; and c) better support for IPv6.

Thanks for @anascko, @vaishnavibhat, @PraveenPenguin and @smruti77

Signed-off-by: Beraldo Leal <bleal@redhat.com>
Signed-off-by: Oksana Vohchana <ovoshcha@redhat.com>
Signed-off-by: Praveen K Pandey <praveen@linux.vnet.ibm.com>
Signed-off-by: Vaishnavi Bhat <vaishnavi@linux.vnet.ibm.com>