utils/cpu: Introduce VENDORS_MAP module's attribute
The get_cpu_vendor_name() uses an internal mapping of
name -> 'string expected to be in /proc/cpuinfo'. In this
change that map became an attribute of avocado.utils.cpu
module, this way 1) users have visibility about the logic
used to determine the vendor's name, and 2) users can also
overwrite VENDORS_MAP if needed.

Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>