nrunner: use current PATH on avocado-runner-exec and avocado-runner-tap
Sometimes, third-party applications and/or tests will call binaries
without a full path. They need a previous path to be set.  When running
an `avocado-runner-exec` or `avocado-runner-tap` we use `kwargs` as
environment variables. But `env` argument from `subprocess.run()` will
pass to the subprocess only the desired dict. This PR fixes that.

Signed-off-by: Beraldo Leal <bleal@redhat.com>