Add avocado plugin to install custom python packages for tests
This patch adds support for installing custom python packages
which are not available on the environment, as provided by users
if enabled through a plugin.

Background, some of the python modules imported by the tests
are not available in default installation of certain environment,
due to which the testcase would fail and these modules(python packages)
can not be installed as a part of test as it would be importing
them as part of test, so only option is to install them before even
starting the test, so this patch tries to add a avocado plugin
to address this issue, by taking the list of python packages from
user through avocado config, it will install them before starting
the job run and optionally it will exit job incase of failure and
optionally it will uninstall packages during end of job run.

Signed-off-by: Satheesh Rajendran <sathnaga@linux.vnet.ibm.com>