nrunner: implement a task Scheduler
The nrunner architecture is based on the execution of independent
tasks, one for each test the user has asked to be run.

The current nrun plugin implements and ad-hoc method of scheduling
these tasks.  The same logic, or a better one, needs to implemented in
the "avocado.plugins.runner_nrunner" module, to make the nrunner
features such as running tests in parallel available to the Avocado
job.

Signed-off-by: Cleber Rosa <crosa@redhat.com>