Plugin Asset: introduces --ignore-errors to assets fetch command.
When the command `avocado assets fetch` fails to download one of
a list of assets, it ends with return code different from 0.

It may be useful to allow the command to end with a 0 return code,
independent of any problem it had during the download execution.
This allows, for example, Travis CI to continue.

Signed-off-by: Willian Rampazzo <willianr@redhat.com>