nrunner: improve the command selection for executable executions
The nrunner has the potential to run a Task (or runnable) in two
different ways:

 1) calling the Python code `for status in task_instance.run(): ...`
 2) executing a command such as `avocado-runner-kind task-run ...`

Currently the second option is favoured and better developed, but the
first option can be useful and has the potential to be much more
lightweight.

This changes the naming of the functions involved in the selection of
the executables for running a task, changes the documentation and
improves the behavior to make sure that the executables declares
formal support for a given type.  This is done by calling the required
"capabilities" command.

Signed-off-by: Cleber Rosa <crosa@redhat.com>