nrunner: do not use a default/last resort runner
There was logic so far to attempt to use the "avocado.core.nrunner"
module as a "default" runner, but in practice that is already
prevented by the "check task requirements" routine.

Also, having a runner being executed for an runnable or task that
doesn't match its intended type is the kind of implicit behavior that
should not be encouraged.

Signed-off-by: Cleber Rosa <crosa@redhat.com>