nrunner: move runner selection to runnable
The selection of a runner was initially introduced as a set of methods
to the Task class.  Most use cases will involve running a Task (which
always contain a Runnable, but not using a Runnable by itself), but
the runner really acts on the Runnable.  This can be evidenced by the
"task.runnable.kind" references in the methods being moved to the
Runnable class.

With this change, it's expected that we will be able to determine if a
runnable has its requirements satisfied.  Right now the only
requirement is really a suitable Runner implementation for the
Runnable kind.

Then, if found to be necessary, we can then implement and determine
extra requirements that are related to a Task that are not necessarily
related to a Runnable.  For instance, the a Task may be configured to
require a specific scheduling/spawing condition, such as to be
executed on a container, which to a Runnable should be transparent (a
Runnable should only care that it's being executed by a suitable
Runner).

Signed-off-by: Cleber Rosa <crosa@redhat.com>