nrunner: move runner selection to runnable
Analogous to the pick_runner_command() method, the pick_runner_class()
method selects the known runner class based on the runnable kind.

There are a few differences from the command selection, given that
for a Python class we don't attempt to find a Python module in the system
and load its code.  For Python classes we depend on a proper and active
registration.

Signed-off-by: Cleber Rosa <crosa@redhat.com>