nrunner: return success on spawn_task()
The current spawner implementations do not signal if they believe that
the spawning was successful or not.

Even though we can not predict and catch all possible errors
conditions, this will allow the spawner to communicate the upper
layers, say an scheduler/job, that one should not wait or check the
"alive" status on a task that failed even to be spawned.

Signed-off-by: Cleber Rosa <crosa@redhat.com>