selftests/unit/test_datadir.py: fix settings_dir_alternate_dynamic
Since be47dcc45f, the data_dir module is supposed to reflect changes
in the configuration dynamically.  But, when that test was ported from
flexmock to mock on cffba44da8, the import of the module without an
alternate configuration was lot (the import was moved to within the
alternate configuration mock block).

This change actually implements what's described in the docstring,
importing the data_dir module with one setting active, and later using
another setting and checking for the altenate values.

Signed-off-by: Cleber Rosa <crosa@redhat.com>