nrunner/spawners: foundation for output file stream
In order to provide a download functionality for output data each spawner needs
to provide such mechanism. The idea is simple: each spawner knows how to
retrieve data generated by their tasks, so we need to delegate the retrieve
functionality to each spawner. This change will implement the basic foundations
for local stream only. If a spawner needs a specific stream method, then it has
to override the 'stream_output' method.

Signed-off-by: Beraldo Leal <bleal@redhat.com>