Spawner: introduce basic tests
The spawners will, for the vast majority of time, be dealing with the
real world, spawning processes, containers (and possibly other
execution environments such as VMs in the future).

Because of the randonmess inherent to the real world, we need a
predictable behavior to test its interfaces, introduce here as as the
MockSpawner.  It's also useful to test how upper layer code will
handle the randomness without actually resorting to running processes,
containers, etc, and for that reason another spawner implementation is
introduced and tested here.

Signed-off-by: Cleber Rosa <crosa@redhat.com>