Process Spawner: add a task to wait on the task process returncode
From the documentation it's not clearly apparent that returncode
needs to be `wait()`ed for explicitly before the returncode is
set.  This means that the current implementation is broken because
ProcessSpawner.is_task_alive() will never return False.

With this change, a task is created in the default event loop that
will attempt wait on the process.  Like all asyncio tasks and other
coroutines, it works in a cooperative way, so other coroutine(s) must
yield to give it a change to run.

This should not be a problem because spawn_task() is also a coroutine,
and it's expected that the layers above this (such as the current
nrunner code, and soon to be introduced scheduler) have a running
loop.

Signed-off-by: Cleber Rosa <crosa@redhat.com>