[PoC / DO NOT MERGE]: migrate nrun to use scheduler
This is intended to be used as "proof" that the scheduler has the
features needed, and a sane API to be used in Avocado.  I picked
nrun because it's clearer to see how it can be applied than if I
were, at this point, to attempt to use it in the "runner_nrunner"
plugin.

Signed-off-by: Cleber Rosa <crosa@redhat.com>