[PoC / DO NOT MERGE]: implement a PYTHON_CLASS SpawnMethod
This uses the same ProcessSpawner, and, if it's found that the
runner for the runnable kind is implemented as a Python class,
it shortcuts its invocation.

The traditional implementation we have so far takes a task, and
performs:

 1) Marshaling of a task into command line, using:
    - Task.pick_runner_command()
    - Task.get_command_args()

 2) Executes a subprocess using the combination of the previous
    command and args from the previous step.

 3) Within the execution of that new process, a Task instance
    is recreated with and executed with:
    - BaseRunnerApp.command_task_run()

This implementation of another SpawnMethod is an optmization
in the sense that most of the code, including a relatively
expensive execution of a Python interpreter from scratch,
is skipped, and the Task has its `run()` pretty much directly
in a new process.

The impact of this change will vary from platform to platform, and
from the type of load.  On my system, when running all Avocado
selftests, before this change:

 $ /usr/bin/time avocado nrun selftests/unit/ selftests/functional/ optional_plugins/*/tests/ selftests/*sh
 ...
 1489.39user 142.99system 3:34.34elapsed 761%CPU (0avgtext+0avgdata 228816maxresident)k
 16inputs+174800outputs (0major+25351591minor)pagefaults 0swaps

And after this change:

 788.08user 97.09system 1:57.58elapsed 752%CPU (0avgtext+0avgdata 219112maxresident)k
 16inputs+174664outputs (0major+16754258minor)pagefaults 0swaps

A ~82% performance improvement.

And when running only the unittests (which are much lighter than
say, the functional tests and linters) before this change:

 $ /usr/bin/time avocado nrun selftests/unit/
 ...
 361.91user 44.11system 0:54.64elapsed 743%CPU (0avgtext+0avgdata 73548maxresident)k
 8inputs+9616outputs (0major+8094617minor)pagefaults 0swaps

And after this change:

 27.43user 13.74system 0:32.96elapsed 124%CPU (0avgtext+0avgdata 73192maxresident)k
 0inputs+9808outputs (0major+2948640minor)pagefaults 0swaps

A ~68% performance improvement.

Jobs on Travis can be found for many Python versions and
architectures, before this change:

 https://travis-ci.org/github/clebergnu/avocado/builds/681756765

And after this change:

 https://travis-ci.org/github/clebergnu/avocado/builds/682049369

Note: some tests will fail to run because they'll share resources
with the parent process.  A number of solutions exist, not limited
to, but including, choosing a different "start methods".

Reference: https://docs.python.org/3.5/library/multiprocessing.html#contexts-and-start-methods
Signed-off-by: Cleber Rosa <crosa@redhat.com>