vmimage: support running on non-UNIX platforms
The currently in use os.uname() function to determine the current
architecture is currently limited to "recent UNIX" systems, according
to Python's docs.  So, on systems without that function, the import of
this module crashes.

Also, instead of calling it multiple times, this calls it only once
(where available) and better documents that aspect.

Signed-off-by: Cleber Rosa <crosa@redhat.com>