Release: add a check for the COPR RPM package version
I've noticed that sometimes the COPR repos are not in sync with the
latest commits, which we want to test in the deployment checks.  This
may be due to builds still running, or repos not synced yet, or any
other infrastructure error.

This is a poor man's script to check that the packages matching the
latest Avocado master branch are available and can be properly
installed before attempting to run the complete deployment checks.

The best solution IMO would still be passing parameters to the
deployment checks playbooks, so that it attempts to install the exact
version being tested.

Signed-off-by: Cleber Rosa <crosa@redhat.com>