Remote/VM/Docker runners: deprecate them
The remote, vm and docker runner plugins (the ones that allows users
to run a job on those environments), all depend on the Fabric3, which
in turn depends on many other Python modules.  Fabric3 is dead with
the Python 3 compatible releases of fabric, but it requires an API
overhaul, very close to rewriting those plugins.

Maintaining those plugins is proving increasingly hard due to version
mismatch failures and bitrot, and the migration effort, given the
current plans of the Avocado project, is not really worth it.  With
the N(ext)Runner architecture, the goal is for *individual* tests to
be able to be executed in remote environments, so some of the
functionality may be brought back in the future, but under a very
different implementation, which may include even different
dependencies.

Reference: https://github.com/avocado-framework/avocado/issues/3799
Signed-off-by: Cleber Rosa <crosa@redhat.com>