python-avocado.spec: don't build deprecated runners
The remote, vm and docker runners have been deprecated on release 79.0,
and they don't actually build on newer distros (because of the Fabric
packages and API changes).

Let's remove them from the spec file in anticipation of their removal.

With that change, aexpect is not required anymore, and the custom
logic to install it in the mock build environment is not necessary.

Signed-off-by: Cleber Rosa <crosa@redhat.com>