sysinfo: fix location of implicit directory creation
When running the `avocado sysinfo` command, because of the current
"./" default, the logic that creates the custom "sysinfo-$timestamp"
directory is skipped, and the pre/post/profile directories are created
within the current working directory.  This is not the originally
intended behavior.

With this new default, users should have the previous experience again
(and a clearer expectation with the more verbose help message).

Signed-off-by: Cleber Rosa <crosa@redhat.com>