Job: allow the selection of the test runner by command line
With the migration to the N(ext)Runner, it's handy to run one-off
avocado jobs on the command line, with the current runner and then
compare it with the nrunner implementation.  This command line switch
makes that possible, while fixing the current configuration key to
match the other configurations related to running tests to be
prefixed with "run.".

Signed-off-by: Cleber Rosa <crosa@redhat.com>