Job: use Python's builtin level logging names instead of custom dict
This basically drops a custom replica of Python's own logging level
mapping, because starting with 3.2 the names of the levels as strings
can be used directly.

A possible small impact for users may the need to change the from
lowercase to uppsercase, but with the correct documentation and notice
on the release notes, it may be better than carrying the special logic
here.

Signed-off-by: Cleber Rosa <crosa@redhat.com>