[RFC] Job: switch from settings to future.settings
This is marked as an RFC because it touches on the delicate question
of whether the implicit registration of some "core" options is a
valid alternative to the defaults that still exist on the code.

This seems to be a "chicken and egg" situation, in which the job
itself needs to log stuff (including its own configuration), but
the log level is either hardcoded or implicitly registered and kept.

Even though it goes agains a number of good practices, including the
Zen of Python, the implicit registration seems to be a better option
because in the end, it will be kept in the resulting configuration
explicitly.

Anyway, this is an RFC, so feedback is welcome.

Signed-off-by: Cleber Rosa <crosa@redhat.com>