Implemented the 'avocado jobs get-output-files' subcommand
This is the basic implementation for a command that will allow users to
download output files generated by the tests. This logic is only valid
if you are using the new Spawner architecture. Please not that this is
still experimental and since that 'nrun' don't have the 'job structure'
yet, you need to mock some files inside the 'data' dir.

Signed-off-by: Beraldo Leal <bleal@redhat.com>