nrunner plugin: implement support for job level timeout
While it's not possible for the current nrunner plugin implementation
to kill tasks, it's possible to abort a job execution between tasks.
This implementation does just that, also adding a missing call to
end the result tracking of tests, so that tests not executed are
listed as SKIPs.

If you have a test that runs for about 1 second, and do:

 $ avocado run --test-runner=nrunner --job-timeout=2s test test test test

It should produce:

 RESULTS    : PASS 2 | ERROR 0 | FAIL 0 | SKIP 2 | WARN 0 | INTERRUPT 0 | CANCEL 0

Signed-off-by: Cleber Rosa <crosa@redhat.com>