sysinfo: remove built in knowledge for specific events
Here event is used to represent a job, a test or any other event
with a beginning and end.  Previously the Sysinfo class would have
knowledge about both job and test start and end hooks.

This is unnecessary, given that different instances will be used
at different points.  Let's make the mechanism abstract and not
really bound to the start or end of job or tests.

Before this, the knowledge about whether the sysinfo is running
as part of a test or as part of a job, was also embedded within
the Sysinfo class, now it's up to its user.

One of the tests, test_logger_test_hook(), were relying on the fact
that the default configuration has "per_test" set to False, and
consequently sysinfo content would not captured normally for tests by
default.

Finally, AVOCADO_TEST_SYSINFODIR will only be exported if the per-test
sysinfo collection is actually enabled, fixing a bug.

Signed-off-by: Cleber Rosa <crosa@redhat.com>