sysinfo: run it within the job, before and after all tests
The goal of the sysinfo pre and post captures, is to attempt to
understand changes to the system before and after tests are run.

But, as it stands, the post "hook" as a "JobPost" implementation, it
runs after the job is completed, so the job itself (such as with the Job
API) won't be able to run it.  Also, results formats rendered after
the tests are finished (such as the HTML report) that use the sysinfo
output, won't have them available.

By making them "JobPreTests" and "JobPostTests" the job itself can
run them, and the rendered results too, without disturbing the goal
of having them run before/after all tests.

Signed-off-by: Cleber Rosa <crosa@redhat.com>