[RFC] Remove per-test sysinfo collection
The current implementation of per-test sysinfo collection is,
well, nasty.  It's called by the test itself, and not by whoever
is running the test.

With the world (including testing) moving to containers, and with the
parallel and very different approach of the Avocado N(ext) Runner,
this feature may be ressurrected, but it feels like it's going to
be completely different.

For instance, if running with a "process" spawner, the pre-test will
probably be done once before spawning a batch of tests (to avoid
changes in the same system caused by tests which are starting to run).
And when running with a "podman" spawner, it makes sense to either
capture the information from the container once, or even keep the
entire container instead?

Signed-off-by: Cleber Rosa <crosa@redhat.com>