avocado.utils.gdb: basic Python 3 fixes
This changes the mode of reading and writing to the GDB process stdout
and stdin to use bytes.  Also, becase a different behavior was found
with regards to the I/O layer in Python 3 (when compared to Python 2),
it's necessary to flush after writing to stdin.

These basic fixes are enough to fix the most basic use cases of the
GDB API, such as the one referenced on issue #3521.

Reference: https://github.com/avocado-framework/avocado/issues/3521
Signed-off-by: Cleber Rosa <crosa@redhat.com>