GDB: update external GDB/mi parser related modules
The spark library is a parser generated library, which has been used
in Python itself until version 3.3, and from which I took an updated
version and curated its update to Avocado's current version.  The same
is true for the GDB/mi parser (built using spark) which I curated and
update based on git://git.rtems.org/rtems-tools.git, which contains a
number of Python 3 fixes.

With those, it's possible to run almost all but one of the GDB
examples tests (that exception has different issues, not related to
this parser).

Signed-off-by: Cleber Rosa <crosa@redhat.com>