Job: make 'core.show' a settings that's always registered
The execution of jobs via the job API is currently broken because the
job does "config.get('core.show')", which has a default currently
registed by the avocado/core/parser.py file, only active when there's
a command line parser (that means only when used as part of "avocado
run".

This registers that option oncce `avocado.core` is loaded, which makes
it effectively a core setting, and the th rest of `avocado.core` can
rely on it.

Signed-off-by: Cleber Rosa <crosa@redhat.com>