Yaml to Mux: split initialization and command line registration
This allows the plugin to be used outside the command line, such as
with the Job API.  If the plugin is installed, then the initialization
will be performed transparently (upon loading of avocado.core).

There's a proposal here for a standard section name, following the
plugin name when it's not necessary to distinguish between different
commands or plugin types.  For instance, in this case, it makes no
sense to distinguish between "list" and "run" because only one will be
active, and the options are really relevant to the plugin
(yaml_to_mux).

Signed-off-by: Cleber Rosa <crosa@redhat.com>