Variants: rename args parameter to config
This "args" parameter would be used when that object came from the
command line arguments only.  It's now the norm to call this object
"config", as it comes from the future.settings module, which takes
into account not only the command line parser but configuration files
and even manually crafted dictionaries.

Signed-off-by: Cleber Rosa <crosa@redhat.com>