Pre release job: workaround to have filtering working
On commit 3d743d717 the name of the section name got added the
command name.  But, there's no command when using the Job API.
This adds "None" to the section name just to get the job working,
and expose the issue that this approach causes.

It also fixes a bug that should have been part of 197422e6, because
the references are always location at "run.references".

Signed-off-by: Cleber Rosa <crosa@redhat.com>