Settings: allow options to be registered by multiple parser
Some options are common across multiple components, such as
"--filter-by-tags", which is used when listing and running tests.

It's awkward to refer to them by different sections just because they
would be *used* in different contexts, but represent the exact same
feature and behavior.

This change makes it possible to register multiple options to the same
section/key, but only if explicitly requested to avoid mistakes.

Signed-off-by: Cleber Rosa <crosa@redhat.com>