Filter by tags: make the option common to different contexts
One current side effect of using command names as a side effect is
that the configuration entry will follow the command, but, on Job API
usages, for instance, there's no command being executed.

One possibility would be to rename those option's sections to be
"job", but this is also misleading because if you're listing tests,
then it's not really a job.

Also, the same feature of filtering test by tags, can be used by
different commands, by they will not overlap because one will either
by listing or running a test at a time.

So, the proposal here is to use the "feature" as the section name, and
allow the same section/key to be controlled by different parsers which
AFAICT will only be used one at time (one will only run either a
"list" or a "run" command) and doesn't require a fake section name on
the Job API.

Signed-off-by: Cleber Rosa <crosa@redhat.com>