avocado.plugins.runner: index reorganization
IMO, it's weird to start a loop with a negative index, just because
later in the loop the index will be incremented and a zero based list
will be referenced.

This is a more conventional loop that hopefully will make the code
more straightforward to read, that is, no questions will arise from
readers.

Signed-off-by: Cleber Rosa <crosa@redhat.com>