TestID: move implementation to its own module
The avocado.core.test module is already pretty large, so a re-org is
good in itself.

Also, it's only the conventional test runner that depends on the
avocado.core.test module, because the INSTRUMENTED test is implemented
there (and all other tests inherit from it).  A nrunner implementation
should not need to import avocado.core.test, but should produce
dependable TestIDs, so this is another reason for the split.

Signed-off-by: Cleber Rosa <crosa@redhat.com>