selftests/unit/test_job.py: fixes for using the nrunner plugin
When setting the "run.test_runner" configuration key to "nrunner",
these unittests for the job class fail to take into account that the
test suite will be comprised of nrunner.Tasks, instead of so called
"test factories".

This change takes that into account and makes the tests succeed.  It's
necessary, though, to manually set the runner to verify the fix.

It's expected that we'll add a full coverage for both test runner
implementations soon, and address the good number of functional tests
failures, but this simple fix resolves the issue with all the
unittests.

Signed-off-by: Cleber Rosa <crosa@redhat.com>