Sysinfo: register options so values from config are respected
The avocado.core.job module has been using future.settings for some
time now, instead of the legacy settings module.

While future.settings is able to read configuration files and merge
their content into the produced dictionary, it checks that the option
has been registered in the first place.

This adds one of the sysinfo configurations that are available on
the configuration file we ship, but that is not being registered.
A final version, if this approach is agreed upon, should include
all option in the configuration file.

Reference: https://github.com/avocado-framework/avocado/issues/3873
Signed-off-by: Cleber Rosa <crosa@redhat.com>