Job API: use "job.run.results" section for result related features
As we organize the Job API, we're choosing the namespaces for features.
One common pattern we have, is that a lot of features are simply organized
under a "run." prefix, because of their legacy implementation in
"avocado/plugins/run.py" or the idea that they were features only
available at "avocado run" time.

The proposal here is to try to think first of the underlying job
level for those features which "run.*", and try to organize them
around the different job phases/features.

For instance, a JSON result (or HTML) is only generated when
a job is run.  So IMO, it makes to organize it as:

 * job - no way to generate a JSON elsewhere
 * run - only when running a job it's possible to have data
         for a JSON result file
 * result - that's the general group of feature, and should
            be shared with HTML, xUNIT, etc

So the section for JSON result plugin, in the contect of generating
it when running a job, becomes "job.run.result.json".

Also, tests for the feature using the underlying job API configuration
are added.  I don't have a clear answer at this point (RFC welcome)
but I believe these can serve as a replacement for much of the
functional tests that run the avocado command line app, because we'd
only be testing the command line parsing, in addition to what's been
tested here.

Signed-off-by: Cleber Rosa <crosa@redhat.com>