Unify "avocado" and "python -m avocado" behavior
We currently have a slightly different behavior with these two
commands, and there's no reason for that.  In fact, I lost a good
number of hours because of nuances between them.

Also, given that we already rely on setuptools to generate the scripts
for most Avocado command line apps, let's do the same for the main
avocado command line app.

Signed-off-by: Cleber Rosa <crosa@redhat.com>