avocado.utils.wait.wait_for: use monotonic time
As explained in the Python docs, and suggested in the referenced
issue, this should make Avocado more reliable wrt times.  Running
tests that play with the clock is not that uncommon, and this could
improve the chances of Avocado not breaking under such conditions.

Reference: https://github.com/avocado-framework/avocado/issues/3565
Signed-off-by: Cleber Rosa <crosa@redhat.com>