N(ext) Runner: improve status messages compatibility for avocado-instrumented
The helper method "prepare_status()" was instroduced to make the status
messages by different runners more uniform.  Here, we use that to send
and initial "started" message that was missing.

Unfortunately, it's not trivial at this time to use for the other
messages, because we're relaying and modifying the "old" Avocado
INSTRUMENTED internal messages, and making it compatible with the
N(ext) Runner expectations.

Anyway, this makes the messages generated by this runner compatible
to the point that these runners are now compatible running within
a job.

Signed-off-by: Cleber Rosa <crosa@redhat.com>