N(ext) Runner: implement entry point based runner selection
If the "avocado.core.nrunner" module is being used by the Avocado Job,
it can benefit from being on a system that has setuptools installed
and can pick runner for runnables from there.

But, if the "nrunner" module is being executed by itself, let's say,
on a remote system, it may not access to setuptools or entrypoints,
so it should still rely on its internal runner registry.

This change allows the Avocado Job to run all ranges of test types
(runnable kinds, really) currently supported.  This is true for
executions without spawners and always local, which currently, in a
job, it's the only way they're done.

Signed-off-by: Cleber Rosa <crosa@redhat.com>