Job API: provide more granular information on errors
As more sophisticated uses of the Job API comes, it's helpful to give
more distinguishable information on errors.  This introduces two new
exceptions, that should are raised during the test suite construction.

Also, the "OptionValidationError" name ties the Job errors to a
command line view, which is the opposite goal of the Job API work.

Additionally, the HTML plugin was adjusted to not attempt to produce a
report when the Job itself errors (and there's really no information
to produce an useful report).

Signed-off-by: Cleber Rosa <crosa@redhat.com>