Job API: remove KeyboardException handler at suite creation time
While catching a KeyboardException is a valid thing to do in an
interactive application, doing so exclusively and in an ad-hoc
way at test suite creation makes no sense.

A more uniform and comprehensive handling of that situation may
be added later, at a broader point within the execution of a job.

Signed-off-by: Cleber Rosa <crosa@redhat.com>